Capturing: Use uint32_t as internal screen settings screen ID.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@75279 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/settings.h
@@ -548,7 +548,7 @@ struct CaptureScreenSettings
 };
 
 /** Map for keeping settings per virtual screen. */
-typedef std::map<unsigned long, CaptureScreenSettings> CaptureScreenMap;
+typedef std::map<uint32_t, CaptureScreenSettings> CaptureScreenMap;
 
 /**
  * NOTE: If you add any fields in here, you must update a) the constructor and b)
